# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FlowVerse 3D HTTP API
  version: 0.1.0
  summary: API para diseño, versionado, validación y simulación de flujos.
  description: |
    Contrato HTTP canónico del MVP. La autenticación utiliza cookies HttpOnly
    opacas. Toda mutación autenticada debe incluir `X-CSRF-Token`. Los errores
    comparten un sobre estable y cada respuesta incluye `X-Request-Id`.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
security:
  - cookieAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Crea un usuario y una sesión.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          $ref: "#/components/responses/Session"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Inicia una sesión.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Rota el refresh token y renueva la sesión.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          $ref: "#/components/responses/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoca la sesión actual.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Sesión revocada.
  /v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Obtiene el usuario de la sesión actual.
      responses:
        "200":
          description: Usuario actual.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: Lista proyectos visibles para el usuario.
      responses:
        "200":
          description: Lista paginada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
    post:
      tags: [Projects]
      operationId: createProject
      summary: Crea un proyecto cuyo owner es el usuario actual.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Proyecto creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      operationId: getProject
      summary: Obtiene un proyecto.
      responses:
        "200":
          description: Proyecto.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Cambia nombre o descripción.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Proyecto actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Elimina lógicamente un proyecto.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Proyecto eliminado.
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/projects/{projectId}/members:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      operationId: listProjectMembers
      summary: Lista miembros y roles.
      responses:
        "200":
          description: Miembros.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      tags: [Projects]
      operationId: addProjectMember
      summary: Añade un usuario registrado como editor o viewer.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Miembro añadido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/projects/{projectId}/flows:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Flows]
      operationId: listFlows
      summary: Lista los flujos del proyecto.
      responses:
        "200":
          description: Flujos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlowSummary"
    post:
      tags: [Flows]
      operationId: createFlow
      summary: Crea un flujo con borrador inicial.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlowRequest"
      responses:
        "201":
          description: Flujo creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowSummary"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/flows/{flowId}:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    get:
      tags: [Flows]
      operationId: getFlow
      summary: Obtiene identidad y estado del flujo.
      responses:
        "200":
          description: Flujo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowSummary"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Flows]
      operationId: updateFlow
      summary: Cambia metadatos permanentes del flujo.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowMetadataInput"
      responses:
        "200":
          description: Flujo actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowSummary"
    delete:
      tags: [Flows]
      operationId: deleteFlow
      summary: Elimina lógicamente el flujo.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Flujo eliminado.
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/flows/{flowId}/draft:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    get:
      tags: [Versions]
      operationId: getFlowDraft
      summary: Obtiene el borrador activo con ETag.
      responses:
        "200":
          description: Documento editable.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "./schemas/flow-definition.schema.json"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Versions]
      operationId: replaceFlowDraft
      summary: Reemplaza el borrador aplicando control optimista.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/flow-definition.schema.json"
      responses:
        "200":
          description: Borrador guardado.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "./schemas/flow-definition.schema.json"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v1/flows/{flowId}/draft/restore:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    post:
      tags: [Versions]
      operationId: restoreFlowDraft
      summary: Restaura una versión publicada como nuevo borrador activo.
      description: |
        Copia la definición inmutable indicada, normaliza el documento y
        actualiza también el nombre y la descripción del flujo. La versión
        debe pertenecer al mismo flujo. `If-Match` protege cambios concurrentes.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreFlowDraftRequest"
      responses:
        "200":
          description: Versión restaurada como borrador editable.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            X-Draft-Revision:
              $ref: "#/components/headers/DraftRevision"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreFlowDraftResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /v1/flows/{flowId}/versions:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    get:
      tags: [Versions]
      operationId: listFlowVersions
      summary: Lista versiones publicadas.
      responses:
        "200":
          description: Versiones ordenadas de reciente a antigua.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlowVersion"
  /v1/flows/{flowId}/diff:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    get:
      tags: [Versions]
      operationId: diffFlowRevisions
      summary: Compara semánticamente dos revisiones de un flujo.
      description: |
        Un selector omitido representa el borrador actual. Debe indicarse al
        menos una versión; esto permite versión→draft, draft→versión o
        versión→versión. El orden de nodos, conexiones y variables no genera
        ruido, y cada cambio se clasifica como visual, behavioral o breaking.
      parameters:
        - name: baseVersionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: targetVersionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Diff determinista entre las dos revisiones.
          content:
            application/json:
              schema:
                $ref: "./schemas/flow-diff.schema.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/flows/{flowId}/copilot:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    post:
      tags: [Copilot]
      operationId: adviseFlowWithEvidence
      summary: Recomienda mejoras citando evidencia verificable del flujo.
      description: |
        Construye localmente un paquete minimizado con validación, análisis,
        estructura y, opcionalmente, diff semántico e incidente. Los valores
        de configuración, inputs, outputs y payloads de eventos no se envían
        al proveedor. El servidor descarta toda sugerencia cuyas citas o
        acciones no puedan resolverse contra ese paquete.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceCopilotRequest"
      responses:
        "200":
          description: Recomendaciones y evidencia resoluble utilizada.
          content:
            application/json:
              schema:
                $ref: "./schemas/evidence-copilot.schema.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/UpstreamFailed"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/flows/{flowId}/publish:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    post:
      tags: [Versions]
      operationId: publishFlowDraft
      summary: Valida y publica una copia inmutable del borrador.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "201":
          description: Versión publicada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowVersion"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v1/flow-versions/{versionId}:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    get:
      tags: [Versions]
      operationId: getFlowVersion
      summary: Obtiene una versión inmutable.
      responses:
        "200":
          description: Versión y documento.
          content:
            application/json:
              schema:
                type: object
                required: [version, definition]
                properties:
                  version:
                    $ref: "#/components/schemas/FlowVersion"
                  definition:
                    $ref: "./schemas/flow-definition.schema.json"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/flow-versions/{versionId}/validate:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    post:
      tags: [Versions]
      operationId: validateFlowVersion
      summary: Ejecuta validación estructural y semántica.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Resultado de validación.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationReport"
  /v1/flow-versions/{versionId}/analysis:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    get:
      tags: [Versions]
      operationId: analyzeFlowVersion
      summary: Calcula métricas estáticas del grafo.
      responses:
        "200":
          description: Análisis.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowAnalysis"
...
  /v1/runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      tags: [Runs]
      operationId: getRun
      summary: Obtiene estado y resumen de una ejecución.
      responses:
        "200":
          description: Ejecución.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowverse_access
...
  schemas:
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 12
          maxLength: 128
        displayName:
          type: string
          minLength: 1
          maxLength: 80
...
  responses:
...
    Session:
      description: Sesión activa; las cookies se establecen con Set-Cookie.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Session"
...
    Unauthorized:
      description: Sesión inexistente, vencida o revocada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Estado actual incompatible con la operación.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationFailed:
      description: Entrada o grafo inválido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...